requirements-dev.txt
flake8
autopep8
pytest
pytest-cov
pytest-asyncio
python-dotenv
mypy
isort

openai==0.27.10
tiktoken==0.4.0
weaviate-client==3.23.2
tenacity==8.2.3